+++ src/components/result/Result.tsx
